Switch elb to use defsec
