+++ internal/app/tfsec/rules/aws/elb/alb_not_public_rule_test.go
@@ -1,4 +1,4 @@
-package elbv2
+package elb
 
 import (
 	"testing"
